+++ apps/react/utils/workbook.ts
@@ -1,6 +1,6 @@
 import { Flatfile } from '@flatfile/api'
 import { sheet } from '@/utils/sheet'
-const sheet2 = {...sheet, name: 'Contacts 2', slug: 'contacts-2'}
+const sheet2 = {...sheet, name: 'Contacts 2', slug: 'contacts2'}
 export const workbook: Flatfile.CreateWorkbookConfig = {
   name: 'Contacts',
   sheets: [sheet, sheet2],
