+++ packages/embedded-utils/src/utils/index.ts
@@ -1,4 +1,5 @@
 export { createWorkbookFromSheet } from './createWorkbookFromSheet'
+export { createWorkbookFromSheets } from './createWorkbookFromSheets'
 export { fetchEventToken, EventSubscriber } from './EventSubscriber'
 export { getErrorMessage } from './getErrorMessage'
 export { initializePubnub } from './initializePubnub'
