+++ packages/react/src/components/index.ts
@@ -1,5 +1,9 @@
+export * from './CombinedWorkbook'
+export * from './EmbeddedIFrameWrapper'
 export * from './FlatfileContext'
 export * from './FlatfileProvider'
-export * from './InitSpace'
-export * from './Space'
+export * from './legacy/InitSpace'
+export * from './legacy/Space'
+export * from './Sheet'
+export * from './SimplifiedWorkbook'
 export * from './Workbook'
