+++ packages/react/src/hooks/index.ts
@@ -1,8 +1,5 @@
 export * from './useFlatfile'
-export * from './useCreateListener'
-export * from './usePortal'
-export * from './useSpaceTrigger'
-export * from './useSpace'
 export * from './useListener'
 export * from './usePlugin'
 export * from './useEvent'
+export * from './legacy'
