+++ packages/react/src/types/iFrameProps.ts
@@ -1,0 +1,15 @@
+import { ISpace } from '@flatfile/embedded-utils'
+
+export type IFrameTypes = Partial<
+  Pick<
+    ISpace,
+    | 'iframeStyles'
+    | 'mountElement'
+    | 'exitText'
+    | 'exitTitle'
+    | 'exitPrimaryButtonText'
+    | 'exitSecondaryButtonText'
+    | 'displayAsModal'
+    | 'closeSpace'
+  >
+>
