+++ packages/react/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './IReactInitSpaceProps'
 export * from './IReactSpaceProps'
+export * from './iFrameProps'
