+++ packages/react/src/utils/initializeSpace.tsx
@@ -4,7 +4,7 @@ import { authenticate } from './authenticate'
 import { getErrorMessage } from '@flatfile/embedded-utils'
 
 export const initializeSpace = async (
-  flatfileOptions: IReactSpaceProps & { sheet?: Flatfile.SheetConfig }
+  flatfileOptions: IReactSpaceProps & { sheets?: Flatfile.SheetConfig }
 ): Promise<Flatfile.SpaceResponse> => {
   let space
   const {
@@ -24,7 +24,7 @@ export const initializeSpace = async (
     labels,
     translationsPath,
     languageOverride,
-    sheet,
+    sheets,
   } = flatfileOptions
 
   try {
@@ -37,7 +37,7 @@ export const initializeSpace = async (
     try {
       space = await limitedAccessApi.spaces.create({
         name,
-        autoConfigure: !workbook && !sheet,
+        autoConfigure: !workbook && !sheets,
         ...spaceBody,
         labels: ['embedded', ...(labels || [])],
         ...(environmentId !== undefined && { environmentId }),
